Hazard: add hazard detection unit
Add hazard detection unit, to insert the stall between
the load and the instruction dependent on it. such as:

0x00 li x10, 100
0x04 li x11, 200
0x08 add x1, x10,x11
0x0c lw x3, (x0)
0x10 add x2, x3, x1

the add instruction at 0x10 will be stalled because it
need data in x3, but the data is not ready, and cannot
be fixed by forwarding logic.

BTW, adding the forward logic in register file, so
read after writen  the same register would get the
correct data.

Signed-off-by: flynn <francisco_flynn@foxmail.com>