pipeline: fix bug with reset signal
we need to hold the instruction until the rst signal is
high, and ignore the control signal. otherwise, if the
first instruction after reset is a branch instruction,
ID_FLUSH signal will be set, so the first instruction
will be flushed, and pc_alu will be set to RESET_PC,
thus the program will enter a dead loop.

such as:

10000000 <_start>:
10000000:	0040006f          	j	10000004 <reset_vector>

10000004 <reset_vector>:
10000004:	51e05073          	csrwi	0x51e,0
10000008:	00000193          	li	x3,0

Signed-off-by: flynn <francisco_flynn@foxmail.com>