mmio: Add mmio in memory stage
let the cpu capable of decode memory mapped I/O, if the
MSB of the address is one, then it represent for a mmio.

Also, implement clock counter and instruction counter
register, which is at 80000010 and 32'h80000014.

we are going to let cpu to control our uart in the next
step.

Signed-off-by: flynn <francisco_flynn@foxmail.com>